SOC-1505 | Abstracting editor logic to base controller and fixes
