+++ front/templates/main/components/discussion-header.hbs
@@ -1,7 +1,7 @@
 {{#if showContent}}
 	<div class="header clearfix background-theme-color">
 		<h1>{{siteName}}</h1>
-		<div {{action 'openEditor'}} class="new-post">
+		<div {{action 'toggleEditor' true}} class="new-post">
 			{{svg 'new-post-icon' viewBox='0 0 56 50'}}
 		</div>
 		<div {{action 'showSortComponent'}} class="sort">
