+++ front/templates/main/discussion/forum.hbs
@@ -1,7 +1,7 @@
 {{discussion-header
 	hideSortComponent='hideSortComponent'
 	showSortComponent='showSortComponent'
-	openEditor='openEditor'
+	toggleEditor='toggleEditor'
 	showContent=true
 	smartBannerVisible=smartBannerVisible
 	sortBy=sortBy
