+++ front/templates/main/discussion/post.hbs
@@ -84,8 +84,7 @@
 					{{#unless model.firstPost.isDeleted}}
 						{{discussion-reply-editor
 							create='create'
-							toggleEditorActive='toggleEditorActive'
-							isActive=editorActive
+							toggleEditor='toggleEditor'
 							isRequesterBlocked=model.isRequesterBlocked
 							replies=model.replies
 							requestErrorMessage=model.errorMessage
